Refactor some of the code
Add helper function swapScreens that applies the rendering class to the desired element that also has the screen class.

Removed unneeded checks in clearBoard to whether classList contains a class.

Refactored clearBoard to be before initGame; addUiListeners to be with the other functions that add listeners; addUiListeners call into initGame function.