Add winning overlay
Winning the game displays the winning overlay, which shows the difficulty completed and time taken. Removed div in html that was used to print game status of winning or answering incorrectly. Overlay has a X button and the background is clickable, allowing the message to be closed.

Pulled out the timer formatting into a getTime function which is now used by both the updateTimer function and the showWinning message function.

TODO: Yet to implement continue button function to start a new game from winning overlay.