Add dropdown menu and reveal puzzle button
Clicking the 3 dots on the toolbar opens up a dropdown menu. Dropdown contains a reveal puzzle button, when clicked closes the dropdown. Puzzle solution is revealed with tiles filled in using blue text, by giving tiles class of revealed. Made sure to remove revealed class on clearboard.

Z-index of dropdown set to 1 to be in front of toolbar. Pushed the z-index of overlay background to be in front of the dropdown at index of 2.

Changed the toolbar border colour to be lighter.