Add puzzle reset function
Second option in dropdown list. Clears the board state and re-populates with the current puzzle to restart the game.

Added event listener to query clicks outside of the dropdown menu. Clicking outside will close the dropdown.

Cleared the tileSelected variable when the board is cleared. Not clearing prevents selectTile from selecting the same tile again.