Add closing animations to overlay cards
Instead of toggling the shown class with the toggleOverlay function, add a close class which plays the animation before removing the show class. This allows the animation to play before setting the display to none.

A timeout is set of 3ms which is about the duration of the animation. The shown class is removed before the animation class is removed. This prevents the animation from resetting while the card is still visible.

Refactored the toggleOverlay class into openOverlay and closeOverlay functions. The closeOverlay function runs the animation on the specified element id.