Update ui
Increase the width and padding for close buttons. This is to increase the hitbox size. One close button was missing the class of button.

Wrapped the timer in a label so that clicking either the pause button or clock would activate the pause screen. Renamed the display-timer div to timer-unit. Set the enclosing label to display-timer. Changed the margins around pause-button and timer-unit to be smaller. No longer uses pause button to register click, but use enclosing display-timer label instead.

Added missing stopTimer call when resetting the board.