Change style of screens
Give the game screens (menu-container, game-container, loading-screen) a display of flex. Center content in the middle of the screen.

Removed margins from body.

Reassigned menu-description class to the subheading of the title. Reassigned the 'Choose your Puzzle:' div to a class of menu-prompt.

TODO: Add icon for the main menu.