Add check cell function
Clicking check cell will check if the guess in the current selected tile is correct. If the answer is correct, gives the tile a revealed class which can no longer be modified, and changes the font color to blue.

Changed the way dropdown items are displayed. The dropdown-content div is given a class of game-over when the game ends, hiding all dropdown-items without the game-over-setting class. The only item with the game-over-setting class is the reset puzzle button.

Removed redundant CSS of game-tile-button.