Hide pause function when game over
Give the display-timer element a class of game-over. The game-over class will cause the timer to stop accepting mouse clicks, and hide the pause button. The game-over class is added when game is over, and removed again when a new game  begins or the game is reset.

Removed line which used hide-setting class that is no longer used.